Test failure is fine
These test results are now different, but no more/less broken.

Previously, the inline parsers would run after the whole source file was
parsed, so errors would be reported at the end of the file.

Now, inline parsers run in their own context, so errors are relative to
the start of the inline block.

Both are wrong of course. We should be able to store the location (row,
col) where the message is generated, and then some arithmetic to get the
correct row/col in the source string.